#169: hard-error out-of-range integer literals (E121) + lossless i128 literals
Out-of-range typed integer literals built clean and silently truncated/wrapped
at codegen (`let a: Int8 = 200` ran as -56; `Int64 = 9223372036854775808`
became Int64.minValue) — char escapes were range-checked but numeric literals
never were.

Two parts:

1. Widen `HirLiteral::Integer(i64)` -> `i128` so every valid magnitude
   round-trips: `UInt64.maxValue` (2^64-1) stays a positive magnitude and the
   `2^63`/`i64::MIN` bit-pattern collision (which hid out-of-range Int64
   literals) can no longer occur. `parse_int` becomes a clean
   `i128::from_str_radix` (no more u64 bit-cast). Pattern matching keeps its
   i64 range arithmetic via a cast at the flat_pat boundary; an out-of-range
   pattern literal is itself diagnosed by E121.

2. New post-inference analyzer `E121 integer_literal_out_of_range` checks each
   integer literal against the fixed-width range of its resolved type
   (Int8..Int64 / UInt8..UInt64 and the lang i8..i64 / u8..u64 intrinsics). A
   leading `-` is the `negate` operator over the literal, so a unary-`negate`
   whose receiver is an integer literal is checked as the negated value — this
   is what makes `Int8.minValue = -128` valid while `-129` is rejected.
   Non-integer / generic / unresolved types are skipped.

Out-of-range literals now fail the build (no binary, non-zero exit); boundary
values (-128, 255, 2^63-1, 2^64-1) still build and run. Full suite 3463 green.

+++ lib/kestrel-hir/src/body.rs
@@ -435,7 +435,13 @@ pub enum HirPat {
 /// HIR literals have been parsed into their concrete types.
 #[derive(Clone, Debug, PartialEq)]
 pub enum HirLiteral {
-    Integer(i64),
+    /// Integer literal magnitude as `i128` so every valid target value
+    /// round-trips losslessly: `UInt64.maxValue` (2^64-1) stays a positive
+    /// magnitude, and the `2^63`/`i64::MIN` collision that hid out-of-range
+    /// `Int64` literals can no longer occur. Negation is applied by the
+    /// `negate` operator, not folded here. The range check lives in the
+    /// integer-literal-range analyzer (E121).
+    Integer(i128),
     Float(f64),
     /// Decoded string value plus any escape-sequence errors discovered during
     /// lowering. Errors are data on the node — a separate analyzer turns them
